Багфікс

+++ updater.ps1
@@ -11,7 +11,9 @@ Set-Location $RootDir;
 . $("$RootDir\\functions.ps1");
 $Message = $XMLConfig.config.messages.unpacking;
 $MhddosPath = $RootDir + "\\" + $XMLConfig.config.folders.load + "\\";
-$null = Remove-Item -Path $MhddosPath -Recurse -Force | Out-Null;
+if (Test-Path $MhddosPath) {
+    $null = Remove-Item -Path $MhddosPath -Recurse -Force | Out-Null;
+}
 Clear-Line $("$Message mhddos_proxy")
 Set-Location $RootDir;
 $Message = $XMLConfig.config.messages.unpacking;
@@ -19,8 +21,12 @@ $LoadURL = $XMLConfig.config.links.load;
 $LoadPath = $RootDir + "\\" + $XMLConfig.config.folders.load + "\\";
 Clear-Line $("$Message load")
 $GitArgs = "clone $LoadURL $LoadPath";
+$GitExe = $("$RootDir\\$GitPath\\bin\\git.exe");
 Start-Process -FilePath $GitExe -ArgumentList $GitArgs -Wait -WindowStyle Hidden;
 
+$PyPath = $XMLConfig.config.folders.python;
+$PythonFolder = $("$RootDir\\$PyPath")
+$PythonExe = $PythonFolder + "\\" + "python.exe";
 Set-Location $LoadPath;
 Clear-Line $("$Message requirements.txt")
 $PyArgs = "-m pip install -r requirements.txt";
